Split Ruleset file to make it pure single-class
