+++ core/src/com/unciv/models/ruleset/Ruleset.kt
@@ -1,16 +1,9 @@
 package com.unciv.models.ruleset
 
-import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.files.FileHandle
 import com.unciv.json.fromJsonFile
 import com.unciv.json.json
 import com.unciv.logic.BackwardCompatibility.updateDeprecations
-import com.unciv.logic.UncivShowableException
-import com.unciv.logic.map.MapParameters
-import com.unciv.models.Counter
-import com.unciv.models.ModConstants
-import com.unciv.models.metadata.BaseRuleset
-import com.unciv.models.metadata.GameParameters
 import com.unciv.models.ruleset.nation.CityStateType
 import com.unciv.models.ruleset.nation.Difficulty
 import com.unciv.models.ruleset.nation.Nation
@@ -22,58 +15,15 @@ import com.unciv.models.ruleset.tile.TileImprovement
 import com.unciv.models.ruleset.tile.TileResource
 import com.unciv.models.ruleset.unique.IHasUniques
 import com.unciv.models.ruleset.unique.Unique
-import com.unciv.models.ruleset.unique.UniqueTarget
 import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.models.ruleset.unit.Promotion
 import com.unciv.models.ruleset.unit.UnitType
 import com.unciv.models.ruleset.validation.RulesetValidator
 import com.unciv.models.stats.INamed
-import com.unciv.models.stats.NamedStats
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.extensions.colorFromRGB
 import com.unciv.utils.Log
-import com.unciv.utils.debug
 import kotlin.collections.set
-
-object ModOptionsConstants {
-    const val diplomaticRelationshipsCannotChange = "Diplomatic relationships cannot change"
-    const val convertGoldToScience = "Can convert gold to science with sliders"
-    const val allowCityStatesSpawnUnits = "Allow City States to spawn with additional units"
-    const val tradeCivIntroductions = "Can trade civilization introductions for [] Gold"
-    const val disableReligion = "Disable religion"
-    const val allowRazeCapital = "Allow raze capital"
-    const val allowRazeHolyCity = "Allow raze holy city"
-}
-
-class ModOptions : IHasUniques {
-    override var name = "ModOptions"
-
-    var isBaseRuleset = false
-    var techsToRemove = HashSet<String>()
-    var buildingsToRemove = HashSet<String>()
-    var unitsToRemove = HashSet<String>()
-    var nationsToRemove = HashSet<String>()
-
-
-    var lastUpdated = ""
-    var modUrl = ""
-    var defaultBranch = "master"
-    var author = ""
-    var modSize = 0
-    var topics = mutableListOf<String>()
-
-    override var uniques = ArrayList<String>()
-
-    @delegate:Transient
-    override val uniqueObjects: List<Unique> by lazy (::uniqueObjectsProvider)
-    @delegate:Transient
-    override val uniqueMap: UniqueMap by lazy(::uniqueMapProvider)
-
-    override fun getUniqueTarget() = UniqueTarget.ModOptions
-
-    val constants = ModConstants()
-}
 
 class Ruleset {
 
@@ -471,176 +421,4 @@ class Ruleset {
     }
 
     fun checkModLinks(tryFixUnknownUniques: Boolean = false) = RulesetValidator(this).getErrorList(tryFixUnknownUniques)
-}
-
-/** Loading mods is expensive, so let's only do it once and
- * save all of the loaded rulesets somewhere for later use
- *  */
-object RulesetCache : HashMap<String,Ruleset>() {
-    /** Whether mod checking allows untyped uniques - set to `false` once all vanilla uniques are converted! */
-    var modCheckerAllowUntypedUniques = true
-
-    /** Similarity below which an untyped unique can be considered a potential misspelling.
-     * Roughly corresponds to the fraction of the Unique placeholder text that can be different/misspelled, but with some extra room for [getRelativeTextDistance] idiosyncrasies. */
-    var uniqueMisspellingThreshold = 0.15 // Tweak as needed. Simple misspellings seem to be around 0.025, so would mostly be caught by 0.05. IMO 0.1 would be good, but raising to 0.15 also seemed to catch what may be an outdated Unique.
-
-
-    /** Returns error lines from loading the rulesets, so we can display the errors to users */
-    fun loadRulesets(consoleMode: Boolean = false, noMods: Boolean = false) :List<String> {
-        val newRulesets = HashMap<String, Ruleset>()
-
-        for (ruleset in BaseRuleset.values()) {
-            val fileName = "jsons/${ruleset.fullName}"
-            val fileHandle =
-                if (consoleMode) FileHandle(fileName)
-                else Gdx.files.internal(fileName)
-            newRulesets[ruleset.fullName] = Ruleset().apply {
-                name = ruleset.fullName
-                load(fileHandle)
-            }
-        }
-        this.putAll(newRulesets)
-
-        val errorLines = ArrayList<String>()
-        if (!noMods){
-            val modsHandles = if (consoleMode) FileHandle("mods").list()
-            else Gdx.files.local("mods").list()
-
-            for (modFolder in modsHandles) {
-                if (modFolder.name().startsWith('.')) continue
-                if (!modFolder.isDirectory) continue
-                try {
-                    val modRuleset = Ruleset()
-                    modRuleset.name = modFolder.name()
-                    modRuleset.load(modFolder.child("jsons"))
-                    modRuleset.folderLocation = modFolder
-                    newRulesets[modRuleset.name] = modRuleset
-                    debug("Mod loaded successfully: %s", modRuleset.name)
-                    if (Log.shouldLog()) {
-                        val modLinksErrors = modRuleset.checkModLinks()
-                        // For extension mods which use references to base ruleset objects, the parameter type
-                        // errors are irrelevant - the checker ran without a base ruleset
-                        val logFilter: (RulesetError) -> Boolean =
-                            if (modRuleset.modOptions.isBaseRuleset) { { it.errorSeverityToReport > RulesetErrorSeverity.WarningOptionsOnly } }
-                            else { { it.errorSeverityToReport > RulesetErrorSeverity.WarningOptionsOnly && !it.text.contains("does not fit parameter type") } }
-                        if (modLinksErrors.any(logFilter)) {
-                            debug("checkModLinks errors: %s", modLinksErrors.getErrorText(logFilter))
-                        }
-                    }
-                } catch (ex: Exception) {
-                    errorLines += "Exception loading mod '${modFolder.name()}':"
-                    errorLines += "  ${ex.localizedMessage}"
-                    errorLines += "  ${ex.cause?.localizedMessage}"
-                }
-            }
-            if (Log.shouldLog()) for (line in errorLines) debug(line)
-        }
-
-        // We save the 'old' cache values until we're ready to replace everything, so that the cache isn't empty while we try to load ruleset files
-        // - this previously lead to "can't find Vanilla ruleset" if the user had a lot of mods and downloaded a new one
-        this.clear()
-        this.putAll(newRulesets)
-
-        return errorLines
-    }
-
-
-    fun getVanillaRuleset() = this[BaseRuleset.Civ_V_Vanilla.fullName]!!.clone() // safeguard, so no-one edits the base ruleset by mistake
-
-    fun getSortedBaseRulesets(): List<String> {
-        val baseRulesets = values
-            .filter { it.modOptions.isBaseRuleset }
-            .map { it.name }
-            .distinct()
-        if (baseRulesets.size < 2) return baseRulesets
-
-        // We sort the base rulesets such that the ones unciv provides are on the top,
-        // and the rest is alphabetically ordered.
-        return baseRulesets.sortedWith(
-            compareBy(
-                { ruleset ->
-                    BaseRuleset.values()
-                        .firstOrNull { br -> br.fullName == ruleset }?.ordinal
-                        ?: BaseRuleset.values().size
-                },
-                { it }
-            )
-        )
-    }
-
-    /** Creates a combined [Ruleset] from a list of mods contained in [parameters]. */
-    fun getComplexRuleset(parameters: MapParameters) =
-        getComplexRuleset(parameters.mods, parameters.baseRuleset)
-
-    /** Creates a combined [Ruleset] from a list of mods contained in [parameters]. */
-    fun getComplexRuleset(parameters: GameParameters) =
-        getComplexRuleset(parameters.mods, parameters.baseRuleset)
-
-    /**
-     * Creates a combined [Ruleset] from a list of mods.
-     * If no baseRuleset is passed in [optionalBaseRuleset] (or a non-existing one), then the vanilla Ruleset is included automatically.
-     * Any mods in the [mods] parameter marked as base ruleset (or not loaded in [RulesetCache]) are ignored.
-     */
-    fun getComplexRuleset(mods: LinkedHashSet<String>, optionalBaseRuleset: String? = null): Ruleset {
-        val baseRuleset =
-                if (containsKey(optionalBaseRuleset) && this[optionalBaseRuleset]!!.modOptions.isBaseRuleset)
-                    this[optionalBaseRuleset]!!
-                else getVanillaRuleset()
-
-        val loadedMods = mods.asSequence()
-            .filter { containsKey(it) }
-            .map { this[it]!! }
-            .filter { !it.modOptions.isBaseRuleset }
-
-        return getComplexRuleset(baseRuleset, loadedMods.asIterable())
-    }
-
-    /**
-     * Creates a combined [Ruleset] from [baseRuleset] and [extensionRulesets] which must only contain non-base rulesets.
-     */
-    fun getComplexRuleset(baseRuleset: Ruleset, extensionRulesets: Iterable<Ruleset>): Ruleset {
-        val newRuleset = Ruleset()
-
-        val loadedMods = extensionRulesets.asSequence() + baseRuleset
-
-        for (mod in loadedMods.sortedByDescending { it.modOptions.isBaseRuleset }) {
-            if (mod.modOptions.isBaseRuleset) {
-                // This is so we don't keep using the base ruleset's uniques *by reference* and add to in ad infinitum
-                newRuleset.modOptions.uniques = ArrayList()
-                newRuleset.modOptions.isBaseRuleset = true
-            }
-            newRuleset.add(mod)
-            newRuleset.mods += mod.name
-        }
-        newRuleset.updateBuildingCosts() // only after we've added all the mods can we calculate the building costs
-
-        return newRuleset
-    }
-
-    /**
-     * Runs [Ruleset.checkModLinks] on a temporary [combined Ruleset][getComplexRuleset] for a list of [mods]
-     */
-    fun checkCombinedModLinks(
-        mods: LinkedHashSet<String>,
-        baseRuleset: String? = null,
-        tryFixUnknownUniques: Boolean = false
-    ): RulesetErrorList {
-        return try {
-            val newRuleset = getComplexRuleset(mods, baseRuleset)
-            newRuleset.modOptions.isBaseRuleset = true // This is so the checkModLinks finds all connections
-            newRuleset.checkModLinks(tryFixUnknownUniques)
-        } catch (ex: UncivShowableException) {
-            // This happens if a building is dependent on a tech not in the base ruleset
-            //  because newRuleset.updateBuildingCosts() in getComplexRuleset() throws an error
-            RulesetErrorList()
-                .apply { add(ex.message, RulesetErrorSeverity.Error) }
-        }
-    }
-
-}
-
-class Specialist: NamedStats() {
-    var color = ArrayList<Int>()
-    val colorObject by lazy { colorFromRGB(color) }
-    var greatPersonPoints = Counter<String>()
 }
